PythonProject/smart_med/requirements.txt
# === Django & REST ===
django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-yasg

django-cors-headers
django-extensions

# === Database ===
psycopg2-binary
python-dotenv
django-environ

# === Async Tasks ===
celery
redis

# === Notifications ===
firebase-admin

# === Image & OCR ===
pillow
pytesseract

# === Utilities ===
requests

pgvector
sentence-transformers
gunicorn